requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Automat==22.10.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
clipspy==1.0.4
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
cryptography==42.0.7
cssselect==1.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.14.0
flatbuffers==24.3.25
fqdn==1.5.1
fsspec==2024.5.0
gast==0.5.4
google-auth==2.29.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
html2text==2024.2.26
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
incremental==22.10.0
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keras==3.3.3
kubernetes==29.0.0
libclang==18.1.1
lxml==5.2.2
lxml_html_clean==0.1.1
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
ollama==0.2.0
onnxruntime==1.18.0
openai==1.30.4
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
posthog==3.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.45
Protego==0.3.1
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.1.0
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytube==15.0.0
PyYAML==6.0.1
pyzmq==26.0.3
queuelib==1.7.0
readability-lxml==0.8.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
Scrapy==2.11.2
Send2Trash==1.8.3
service-identity==24.1.0
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.16.0
tinycss2==1.3.0
tldextract==5.1.2
tokenizers==0.19.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
Twisted==24.3.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
w3lib==2.1.2
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
zipp==3.19.0
zope.interface==6.4.post2
